./fuzz now takes advantage of several cores
Because of reasons, AFL seems not to be aware that several
cores are available. We work around this by manually specifying
`-d X` parameter, where X is the core we want AFL to run.

By default, X will be the number of cores in the machine
according the `nproc`.

`./fuzz 1` is identical to the old `./fuzz primary`.
`./fuzz N` spins up 1 master and N-1 secondaries.

Also, secondaries are set up according to AFL's readme good practices
so some use `-Z` and all run in differnt power modes