Added firmware static analysis to CI (#165)
- Unified the signer and ui static analysis scripts into a single script
is located under the more specific folder ledger/static-analysis. Also
modified the directory of the generated reports to be under the
ledger/static-analysis folder.

- Added a new action to CI that runs static analysis on the firmware code.
A failure in static analysis won't prevent the checks from passing, this
is only informational at the moment. The generated reports are uploaded
as a github artifact and can be analyzed offline.