Adds "screen saver" to signer app
The screen saver screen is just a completely black screen with no active pixels. This
commit introduces the logic to switch to this "screen saver" screen after 30 seconds
of inactivity in the Signer app. When any of the physical buttons is pressed, the usual
"Signer running..." screen is presented again.