Change block size type from size_t to uint8_t
sest_read returns a uint8_t size, which is compared with nvm_block_t.size whenever nvm_load is called. Allowing the caller to register a block with a size over the maximum uint8_t size could lead to an inconsistency when nvm_load is used.

This commit changes the block size to uint8_t to make it consistent.