Move oeedger8r generated sources to their own directory
These sources are not versioned, and are generated during the build. Having them in the same directory as the other source files was causing an inconsistency on the reproducible builds, since the TRUSTED_SRC / UNTRUSTED_SRC variables would expand to different values depending on whether or not a build had already been performed.

This is why running a `make clean` followed by a `make build` would result in a different digest / mrenclave pair compared to running `make clean build` on a directory that already contained these sources.