Add sgx service container to sgx_prv group
The container needs to be part of the sgx_prv group, otherwise it won't be
able to read /dev/sgx_provision

NOTE: the reason why this used to work without setting the group was that
we would perform the attestation step of the onboarding while running the
powhsm binary directly from the host machine with sudo. After changing the
setup to perform these steps against the running service (which runs inside a
docker container), we now need that container to be part of the `sgx_prv`
group.