Fixes static analysis finding for svarint.c
This fixes the following report:

--
Result of operation is garbage or undefined
The result of the left shift is undefined due to shifting by '32', which is greater or equal to the width of type 'int'
--

This finding is flagging that when ctx->offset > 3, but buf[i] == 0, we might attempt to perform a 32-bit shift to `buf[i]`, which is undefined behavior.

The logic has been rewritten to keep the same behavior while avoiding this invalid shift operation.