Silence "dead assignment" warning on hsm_process_apdu
This silences the following finding from the static analyzer:

--
Dead assignment

Value stored to 'tx' is never read
--

This is actually a false positive, since the value of tx is clearly used on hsm_process_exception right below. The problem here is that the THROW macro resolves to a `longjmp` to the begining of the function, which tricks the static analyzer into thinking the value of tx will never be used.

The solution to silence this warning is removing the THROW from the success path and handling the exception inside the FINALLY block. With this pattern, the success code path is linear and won't cause any longjmps, making it clear that the value of tx assigned on line 324 is used on line 330.