Fixes uninitialized variable: bool gathered_settings
Fixes the following finding from static analysis:

```
Branch condition evaluates to a garbage value
```

This happens because we evaluate the value of `gathered_settings` in the error handling logic on line 186, but this variable will only be initialized on line 147. There are 3 jumps to generate_evidence_error before that, in which case we use that uninitialized variable in the branch condition.